address comments and lots of cleanup
