+++ packages/pages/src/dev/server/middleware/serverRenderRoute.ts
@@ -1,18 +1,16 @@
-import { PageLoaderResult } from "../ssr/pageLoader.js";
 import { RequestHandler } from "express-serve-static-core";
 import { ViteDevServer } from "vite";
-import { pageLoader } from "../ssr/pageLoader.js";
+import { propsLoader } from "../ssr/propsLoader.js";
 import { urlToFeature } from "../ssr/urlToFeature.js";
 import page404 from "../public/404.js";
-import { convertTemplateConfigInternalToFeaturesConfig } from "../../../common/src/feature/features.js";
-import { validateTemplateModuleInternal } from "../../../common/src/template/internal/validateTemplateModuleInternal.js";
 import { featureNameToTemplateModuleInternal } from "../ssr/featureNameToTemplateModuleInternal.js";
 import {
   renderHeadConfigToString,
   getLang,
 } from "../../../common/src/template/head.js";
 import { ProjectStructure } from "../../../common/src/project/structure.js";
-import { TEMPLATE_PATH } from "../ssr/constants.js";
+import { TemplateModuleInternal } from "../../../common/src/template/internal/types.js";
+import templateBase from "../public/templateBase.js";
 
 type Props = {
   vite: ViteDevServer;
@@ -39,22 +37,10 @@ export const serverRenderRoute =
         return res.status(404).end(page404);
       }
 
-      validateTemplateModuleInternal(templateModuleInternal);
-
-      const featuresConfig = convertTemplateConfigInternalToFeaturesConfig(
-        templateModuleInternal.config
-      );
-
-      const React = await import("react");
-      const ReactDOMServer = await import("react-dom/server");
-
-      const { template, props }: PageLoaderResult = await pageLoader({
-        url: url.pathname,
-        vite,
-        templateFilename: templateModuleInternal.filename,
+      const props = await propsLoader({
+        templateModuleInternal,
         entityId,
         locale,
-        featuresConfig,
         dynamicGenerateData,
         projectStructure,
       });
@@ -62,28 +48,26 @@ export const serverRenderRoute =
       if (templateModuleInternal.render) {
         res
           .status(200)
-          .set({ "Content-Type": "text/html" })
+          .type(getContentType(templateModuleInternal, props))
           .end(templateModuleInternal.render(props));
         return;
       }
 
-      if (!templateModuleInternal.default) {
-        throw Error(
-          "Either a default export or render function is required in template: " +
-            `/${TEMPLATE_PATH}/${templateModuleInternal.filename}`
-        );
-      }
-      // render the component to its html
-      // Since we are on the server using plain TS, and outside
-      // of Vite, we are not using JSX here
+      const React = await import("react");
+      const ReactDOMServer = await import("react-dom/server");
+
       const appHtml = ReactDOMServer.renderToString(
-        React.createElement(templateModuleInternal.default, props)
+        React.createElement(templateModuleInternal.default || "", props)
       );
 
       const headConfig = templateModuleInternal.getHeadConfig
         ? templateModuleInternal.getHeadConfig(props)
         : undefined;
-      const lang = getLang(headConfig, props);
+
+      const template = await vite.transformIndexHtml(
+        url.pathname,
+        templateBase
+      );
 
       // Inject the app-rendered HTML into the template. Only invoke the users headFunction
       // if they are rendering by way of a default export and not a custom render function.
@@ -93,21 +77,34 @@ export const serverRenderRoute =
             <script type="text/javascript">
               window._RSS_PROPS_ = ${JSON.stringify(props)};
               window._RSS_TEMPLATE_ = '${templateModuleInternal.filename}';
-              window._RSS_LANG_ = '${lang}';
+              window._RSS_LANG_ = '${getLang(headConfig, props)}';
             </script>
-            ${
-              !templateModuleInternal.render && headConfig
-                ? renderHeadConfigToString(headConfig)
-                : ""
-            }
+            ${headConfig ? renderHeadConfigToString(headConfig) : ""}
           </head>`
       );
 
       // Send the rendered HTML back.
-      res.status(200).set({ "Content-Type": "text/html" }).end(html);
+      res
+        .status(200)
+        .type(getContentType(templateModuleInternal, props))
+        .end(html);
     } catch (e: any) {
       // If an error is caught, calling next with the error will invoke
       // our error handling middleware which will then handle it.
       next(e);
     }
   };
+
+/**
+ * Returns the content type based on the getPath's extension. Falls back to 'text/html'.
+ */
+const getContentType = (
+  templateModuleInternal: TemplateModuleInternal<any, any>,
+  props: any
+) => {
+  // TODO: once custom headers are supported at the template level use that instead,
+  // with a fallback to the current logic.
+  const path = templateModuleInternal.getPath(props);
+  const ext = path.includes(".") ? path.split(".").pop() : "";
+  return ext || "text/html";
+};
