+++ packages/pages/src/dev/server/public/templateBase.html
@@ -7,7 +7,7 @@
       import ReactDOM from "react-dom";
       import React from "react";
       const hydrate = async () => {
-        const templates = import.meta.glob("/src/templates/*.(jsx|tsx|ts|js)");
+        const templates = import.meta.glob("/src/templates/*.(jsx|tsx)");
         const routes = Object.keys(templates).map((path) => {
           return {
             name: path.split("/").pop()?.split(".")[0] || "index",
@@ -26,15 +26,8 @@
             throw new Error("Function not implemented.");
           },
         };
-        const { default: Component, render } = await template.getComponent();
-        if (render) {
-          return;
-        }
+        const { default: Component } = await template.getComponent();
         if (!Component) {
-          console.error(
-            "Either a default export or render function is required in template: " +
-              template.path
-          );
           return;
         }
         ReactDOM.hydrate(
